feat: news data from db
