+++ src/app/@modal/default.js
@@ -1,0 +1,3 @@
+export default function Default() {
+  return null;
+}
