+++ src/app/news/page.jsx
@@ -1,57 +1,8 @@
-"use client";
-import news from "@/data/news";
-import NewsCard from "@/components/NewsCard";
-import formatDate from "@/utils/formatDate";
-import FlowSvg from "@/components/FlowSvg";
-import { useRef } from "react";
-import gsap from "gsap";
-import { useGSAP } from "@gsap/react";
+import { db } from "@/utils/firebase/init";
+import { getNews } from "@/utils/firebase/firestore";
+import NewsList from "@/components/NewsList";
 
-gsap.registerPlugin(useGSAP);
-
-export default function News() {
-  const gsapContainer = useRef();
-  const newsCardsRef = useRef([]);
-
-  useGSAP(
-    () => {
-      newsCardsRef.current.forEach((el) =>
-        gsap.fromTo(
-          el,
-          {
-            opacity: 0,
-            y: -100
-          },
-          {
-            opacity: 1,
-            y: 0,
-          }
-        )
-      );
-    },
-    { scope: gsapContainer }
-  );
-
-  return (
-    <main className="main" ref={gsapContainer}>
-      <section className="news news--full">
-        <FlowSvg />
-        <h2 className="title title--center">
-          <span className="title__deco">最新消息</span>
-        </h2>
-        <div className="news__grid">
-          {news.map((news, i) => {
-            return (
-              <NewsCard
-                ref={(el) => (newsCardsRef.current[i] = el)}
-                key={news.date}
-                date={formatDate(news.date)}
-                title={news.title}
-              />
-            );
-          })}
-        </div>
-      </section>
-    </main>
-  );
+export default async function News() {
+  const allNews = await getNews(db);
+  return <NewsList allNews={allNews} />;
 }
