--- src/data/news.js
@@ -1,28 +1,0 @@
-const news = [
-  {
-    date: "2024-04-20T08:55:00.000Z",
-    title: "銀髮族膝蓋訓練講座",
-    content:
-      "保養膝關節對長者而言是很重要的課題。本所將於 4/28 （日）舉辦專為銀髮族設計的膝蓋保健與運動講座，由苑院長與孔治療師現場親自為大家深入淺出分析日常生活容易遇到的膝蓋問題與原因，並提供實用、正確的運動訓練技巧，幫助長者延緩肌力退化、遠離疼痛與保持活力。後續還有趣味問答活動，參加者皆有機會獲得精美小禮品！有興趣的朋友歡迎透過治療所電話或社群連結洽詢詳細資訊。"
-  },
-  {
-    date: "2024-04-03T10:00:01.000Z",
-    title: "臨時休診公告",
-    content:
-      "因地震造成設備與管路損壞，本所將於今日與下周一 (4/8) 臨時休診，以利進行設備與牆面整修作業。已預約之個案我們將協助更改時間，如果您有任何疑問或需要進一步的協助，歡迎透過官方 line、fb、ig 與我們聯繫。造成不便，敬請見諒。"
-  },
-  {
-    date: "2024-03-12T03:00:27.000Z",
-    title: "免費核心運動講座",
-    content:
-      "你聽過核心運動嗎？想鍛鍊核心肌肉卻不得要領嗎？本所將於 3/23 （六）舉辦核心運動講座，由鄭治療師與各位從認識核心肌群開始，逐步進入運動訓練，並點出常見錯誤與注意事項。相信不論是初學者或是健身愛好者都能從中獲益。有興趣的朋友歡迎透過治療所電話或社群連結洽詢詳細資訊。"
-  },
-  {
-    date: "2024-02-01T08:00:01.000Z",
-    title: "春節假期門診時間",
-    content:
-      "新年快樂，龍年行大運！本治療所春節期間門診時間如下：2/9（五）至 2/14（三）休診；2/15（四）起恢復正常營業時間（10:00 - 20:00）。若您於此期間有任何疑問，可以透過官方 line、fb、ig 留言，我們將會在假期結束後盡快回覆，感謝您的支持與理解。"
-  }
-];
-
-export default news;
