+++ src/styles/components/_index.scss
@@ -9,3 +9,4 @@
 @forward "wave";
 @forward "body";
 @forward "handsvg";
+@forward "modal";
