+++ src/styles/components/_modal.scss
@@ -1,0 +1,24 @@
+@use "./../abstracts" as *;
+
+.modal__backdrop {
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  inset: 0;
+  z-index: 22;
+  background-color: rgba($color-secondary-900, 0.5);
+  backdrop-filter: blur(4px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.modal {
+  width: 50%;
+  max-height: 500px;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  overflow: auto;
+}
